Changed years of exp from number to decimal

+++ backend/src/main/java/com/avensys/CVparserApplication/openai/ChatGPTMappedDTO.java
@@ -11,7 +11,7 @@ public class ChatGPTMappedDTO {
 
     private String mobile;
 
-    private int yearsOfExperience;
+    private double yearsOfExperience;
 
     private String[] skills;
 
@@ -56,11 +56,11 @@ public class ChatGPTMappedDTO {
         this.mobile = mobile;
     }
 
-    public int getYearsOfExperience() {
+    public double getYearsOfExperience() {
         return yearsOfExperience;
     }
 
-    public void setYearsOfExperience(int yearsOfExperience) {
+    public void setYearsOfExperience(double yearsOfExperience) {
         this.yearsOfExperience = yearsOfExperience;
     }
